requirements.txt
Django==1.9.4
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.3.3
